done hw#7
