+++ index.html
@@ -58,7 +58,7 @@
         <div class="wall-left"></div>
         <div class="wall-right"></div>
     </div>
-    <script src="main.js" defer></script>
-    <script src="logs.js" defer></script>
+    <script src="main.js"   type = "module" defer></script>
+    <script src="logs.js"  defer></script>
 </body>
 </html>
