+++ utils.js
@@ -1,0 +1,3 @@
+let getRandom = (num) => Math.ceil(Math.random() * num);
+
+export default getRandom;
